Flyttat lite filer och lagt till en klass Project

+++ source/Water/Water3D.h
@@ -43,7 +43,7 @@ private:
     void drawCellTypes();
     void drawFaceVelocities();
     void drawVoxel(double x,double y,double z, double dx,double dy,double dz);
-    void drawSphere(float r, int segs);
+    void drawSphere(double r, int segs);
 public:
     Water3D(int dim = 15);
     void draw();
